Moved non-override fields from EffectItem to EffectItemExtra
This way OBME::EffectItem remains the same size as ::EffectItem, which
is important for compatibility because (1) effect items seem like something
that many 3rd party plugins will want to create, and (2) there is no
game/CS code for allocating them that OBME can patch.
This does increase the size of most effect items by a factor of ~3.